voters,contestants are notified when removed from any election

+++ php/accept_reject_request.php
@@ -2,6 +2,7 @@
 //get user_id and election_id
 require_once("function.php");
 include_once('connection.php');
+include_once('database.php');
 $id=$_POST["id"];
 $id= (explode(" ",$id)[1]);
 $sender_id= (explode("_",$id)[0]);
@@ -11,7 +12,30 @@ $election_id=(explode("_",$id)[1]);
 $starting_date_query= "SELECT election_start_date,election_time_from FROM election WHERE election_id='$election_id'";
 $starting_date= mysqli_fetch_row(mysqli_query($connection2,$starting_date_query))[0];
 $starting_time=mysqli_fetch_row(mysqli_query($connection2,$starting_date_query))[1];
-if(!concluded($starting_date,$starting_time,0) ){
+if(!concluded($starting_date,$starting_time,0) && isset($_POST)){
+    //get needed election details in case you need to send mail
+    $admin_query = "SELECT
+                          election.election_name,users.fname AS admin_fname,users.lname AS admin_lname
+                FROM
+                          election
+                LEFT JOIN
+                          users
+                ON
+                          election.user_id = users.user_id
+                WHERE
+                          election_id = $election_id";
+    $admin = $connection1->prepare($admin_query);
+    $admin->execute();
+    $admin->setFetchMode(PDO::FETCH_ASSOC);
+    $admin = $admin->fetchAll()[0];
+    $election_name = $admin['election_name'];
+    $sender_name = strtoupper($admin['admin_fname'])." ".$admin['admin_lname'];
+
+    //get recipient
+    $recipient = getAllMembers('users',['*'],['user_id','=',$sender_id])[0];
+    $recipient_address = $recipient['email'];
+    $recipient_name = strtoupper($recipient['fname'])." ".$recipient['lname'];
+    $mail_subject = "Your request to join ".$election_name."has been ";
 
     //delete request
     $delete_request_query="DELETE FROM request WHERE user_id='$sender_id' AND election_id='$election_id'";
@@ -19,20 +43,43 @@ if(!concluded($starting_date,$starting_time,0) ){
         if($_POST["action"]==="accept"){
 
             //check if user has not been added to the election before
-            $check_status_query="SELECT * FROM joined WHERE election_id='$election_id' AND user_id='$sender_id'";
-            $check_status=mysqli_fetch_row(mysqli_query($connection2,$check_status_query));
-
-            if(count($check_status)===0){
-
+            if(attached('joined',$sender_id,$election_id) != 'joined'){
                 //add sender to joined
                 $adding_sender_query="INSERT INTO joined(user_id,election_id) VALUES ('$sender_id','$election_id')";
                 if(mysqli_query($connection2,$adding_sender_query)){
+                    //send notification
+                    $mail_subject .= "granted.";
+                    $mail_body = "Hello ".$recipient['username'].".<br>
+                                                This is to notify you that ".$sender_name."; the administrator of ".$election_name."
+                                                has granted your request to join ".$election_name.". Therefore, you are
+                                                now a valid voter in the election. <a href='evoting.oauife.edu.ng'>Login into your account.</a>
+                                                now to view the latest about the election.<br><br> We will like to remind
+                                                you that, at <a href='evoting.oauife.edu.ng'>OAU E-voting system</a>, it
+                                                is our responsibility to provide a reliable and trustworthy one-man-one-vote
+                                                online voting system for you always. Thank you.";
+                    sendEmail($recipient_address,$recipient_name,$mail_subject,$mail_body);
                     echo 'user has been successfully added to election.';
                 }
             }
 
 
         }elseif($_POST["action"]=="reject"){
+            //send notification
+            $mail_subject .= "rejected.";
+            $mail_body = "Hello ".$recipient['username'].".<br>
+                                                This is to notify you that ".$sender_name."; the administrator of ".$election_name."
+                                                has rejected your request to join ".$election_name.". The possible cause
+                                                of this might be either you are not meant to be a voter in this election
+                                                or you did not meet-up to the necessary requirements, as stated by the
+                                                administrator of this election, needed to be a valid voter in this election.
+                                                Whatever the case may be, the administrators best understands why your
+                                                request was rejected.<br><br> We will like to remind you that, at
+                                                <a href='evoting.oauife.edu.ng'>OAU E-voting system</a>, it is our
+                                                responsibility to provide a reliable and trustworthy one-man-one-vote
+                                                online voting system for you always. Thank you.
+                                                <a href='evoting.oauife.edu.ng'>Login into your account.</a>";
+            sendEmail($recipient_address,$recipient_name,$mail_subject,$mail_body);
+
             echo 'Request successfully rejected.';
         }
     }
